+++ kitchen-sink/svelte/src/pages/preloader.svelte
@@ -23,61 +23,62 @@
 
 <!-- svelte-ignore a11y-missing-attribute -->
 <Page>
-  <Navbar title="Preloader" backLink="Back"></Navbar>
+  <Navbar title="Preloader" backLink="Back" />
   <Block>
-    <p>How about an activity indicator? Framework7 has a nice one. The F7 Preloader is made with SVG and animated with CSS so it can be easily resized.</p>
+    <p>
+      How about an activity indicator? Framework7 has a nice one. The F7 Preloader is made with SVG
+      and animated with CSS so it can be easily resized.
+    </p>
   </Block>
 
   <BlockTitle>Default</BlockTitle>
   <Block strong class="row demo-preloaders align-items-stretch text-align-center">
     <Col>
-      <Preloader></Preloader>
+      <Preloader />
     </Col>
     <Col style="background: #000">
-      <Preloader color="white"></Preloader>
+      <Preloader color="white" />
     </Col>
     <Col>
-      <Preloader size={42}></Preloader>
+      <Preloader size={42} />
     </Col>
     <Col style="background: #000">
-      <Preloader size={42} color="white"></Preloader>
+      <Preloader size={42} color="white" />
     </Col>
   </Block>
 
   <BlockTitle>Color Preloaders</BlockTitle>
   <Block strong class="row text-align-center">
     <Col>
-      <Preloader color="red"></Preloader>
+      <Preloader color="red" />
     </Col>
     <Col>
-      <Preloader color="green"></Preloader>
+      <Preloader color="green" />
     </Col>
     <Col>
-      <Preloader color="orange"></Preloader>
+      <Preloader color="orange" />
     </Col>
     <Col>
-      <Preloader color="blue"></Preloader>
+      <Preloader color="blue" />
     </Col>
   </Block>
 
-  <BlockTitle>Multi-color (MD-theme only)</BlockTitle>
+  <BlockTitle>Multi-color</BlockTitle>
   <Block strong class="text-align-center">
-    <Preloader color="multi"></Preloader>
+    <Preloader color="multi" />
   </Block>
 
   <BlockTitle>Preloader Modals</BlockTitle>
   <Block strong>
     <p>With <b>app.preloader.show()</b> you can show small overlay with preloader indicator.</p>
-    <p>
-      <a class="button button-fill" on:click={openIndicator}>Open Small Indicator</a>
-    </p>
+    <p><a class="button button-fill" on:click={openIndicator}>Open Small Indicator</a></p>
     <p>With <b>app.dialog.preloader()</b> you can show dialog modal with preloader indicator.</p>
+    <p><a class="button button-fill" on:click={openDialog}>Open Dialog Preloader</a></p>
     <p>
-      <a class="button button-fill" on:click={openDialog}>Open Dialog Preloader</a>
+      With
+      <b>app.dialog.preloader('My text...')</b>
+      you can show dialog preloader modal with custom title.
     </p>
-    <p>With <b>app.dialog.preloader('My text...')</b> you can show dialog preloader modal with custom title.</p>
-    <p>
-      <a class="button button-fill" on:click={openCustomDialog}>Open Dialog Preloader</a>
-    </p>
+    <p><a class="button button-fill" on:click={openCustomDialog}>Open Dialog Preloader</a></p>
   </Block>
 </Page>